Fix warnings
git-svn-id: http://svn.tinodidriksen.com/tdc/oss/spellers@98 52bd5d93-d692-4aa0-9d28-25f7cffc0c32

+++ mozilla/native/shim.cpp
@@ -114,7 +114,7 @@ bool checkValidWord(const std::string& word, size_t suggs = 0) {
 
 #ifdef _WIN32
 	DWORD bytes = 0, bytes_read = 0;
-	if (!WriteFile(g_hChildStd_IN_Wr, cbuffer.c_str(), cbuffer.size(), &bytes, 0) || bytes != cbuffer.size()) {
+	if (!WriteFile(g_hChildStd_IN_Wr, cbuffer.c_str(), static_cast<DWORD>(cbuffer.size()), &bytes, 0) || bytes != cbuffer.size()) {
 		showLastError("checkValidWord WriteFile");
 		return false;
 	}
